+++ Apps/Components/CMC/2001 - Service/98 - Catcher/clsCatcher.vb
@@ -1,7 +1,7 @@
-﻿Imports System.ComponentModel.Composition.Hosting
-Imports System.Runtime.Versioning
-
-Namespace Ladybug.Log
+﻿Namespace Ladybug.Log
+    ''' <summary>
+    ''' Fields structure is used to store various properties related to logging, such as the sender of the log, the type of fault, the message, the error number, the query, the application version, and other settings.
+    ''' </summary>
     Public Structure Fields
         Public Property FromSender As String
         Public Property TypeOfFaulty As TypeOfFaulties
@@ -14,6 +14,9 @@ Namespace Ladybug.Log
         Public Property ResumeNext As Boolean
         Public Property InternalStackTrace As String
 
+        ''' <summary>
+        ''' TypeOfFaulties enum is used to categorize the type of fault that occurred in the application, such as support service database engine, support service SOAP, support service web, or application runtime.
+        ''' </summary>
         Public Enum TypeOfFaulties
             None = 0
             SupportServiceDatabaseEngine = 20010102
@@ -24,6 +27,9 @@ Namespace Ladybug.Log
 
         Public Property TypeOfLog As TypeOfLogs
 
+        ''' <summary>
+        ''' TypeOfLogs enum is used to categorize the type of log that is being recorded, such as information, warning, fatal, error, or debug.
+        ''' </summary>
         Public Enum TypeOfLogs
             Information
             Warning
@@ -31,20 +37,24 @@ Namespace Ladybug.Log
             [Error]
             [Debug]
         End Enum
+
         Public Property SaveInBetterLog As Boolean
     End Structure
 
+    ''' <summary>
+    ''' Events class is used to handle events related to logging, such as displaying the error reporting form and sending logs to the appropriate destination.
+    ''' </summary>
     Public Class Events
 
         Public WithEvents FRMerc As New CMCv.UI.Canvas.FRMerrorreporting
         Private ReadOnly clsDBsqlite As Database.Engine.SQLiteV3
 
         ''' <summary>
-        ''' Display Error Reporting Form
+        ''' Initializes a new instance of the Events class with the specified SQLite database engine.
         ''' </summary>
-        ''' <remarks></remarks>
-        <SupportedOSPlatform("windows")>
-        Public Sub ShowData(ByVal proLog As Ladybug.Log.Fields)
+        ''' <param name="proLog"></param>
+        <System.Runtime.Versioning.SupportedOSPlatform("windows")>
+        Public Sub ShowData(proLog As Ladybug.Log.Fields)
             Dim varMessage As String
 
             If Not (proLog.ShowErrorReporting) Then
