+++ Apps/Components/LibApp/010. Class/001. Tables Colums Parameters/tDepartment.vb
@@ -3,11 +3,13 @@
         Public ReadOnly TableName As String = "man_department"
         Public ReadOnly S As String = "dep"
         Public ReadOnly C_DepartmentId As String = "department_id"
+        Public ReadOnly C_DepartmentCompany As String = "department_company"
         Public ReadOnly C_DepartmentCode As String = "department_code"
         Public ReadOnly C_DepartmentName As String = "department_name"
+        Public ReadOnly C_DepartmentDescription As String = "department_description"
         Public ReadOnly P_DepartmentId As String = "@DepartmentId"
         Public ReadOnly P_DepartmentCode As String = "@DepartmentCode"
         Public ReadOnly P_DepartmentName As String = "@DepartmentName"
-        'Public ReadOnly 
+        Public ReadOnly P_DepartmentDescription As String = "@DepartmentDescription"
     End Class
 End Namespace
