+++ Apps/Components/LibApp/010. Class/001. Tables Colums Parameters/tPosition.vb
@@ -3,10 +3,14 @@
         ReadOnly Property TableName As String = "man_position"
         ReadOnly Property S As String = "pos"
         ReadOnly Property C_PositionId As String = "position_id"
+        ReadOnly Property C_PositionDepartment As String = "position_department"
         ReadOnly Property C_PositionCode As String = "position_code"
         ReadOnly Property C_PositionName As String = "position_name"
+        ReadOnly Property C_PositionDescription As String = "position_description"
         ReadOnly Property P_PositionId As String = "@PositionId"
+        ReadOnly Property P_PositionDepartment As String = "@PositionDepartment"
         ReadOnly Property P_PositionCode As String = "@PositionCode"
         ReadOnly Property P_PositionName As String = "@PositionName"
+        ReadOnly Property P_PositionDescription As String = "@PositionDescription"
     End Class
 End Namespace
