+++ Apps/Core/Vb/030. Form/020. Commands/004. Management/02. Master/CCIN - Companies/FRMccinEditor.vb
@@ -1,25 +1,17 @@
-﻿Imports System.Runtime.Versioning
-
-Namespace UI.Canvas
+﻿Namespace UI.Canvas
     Public Class FRMccinEditor
 
 #Region "Declaration"
         Public Event EventRecordSaved()
-        Private varDataParameters As New Dictionary(Of String, Object)
-        Private Const tCompany As String = "man_company"
 
-        'Parameters
-        Private Const pCompanyCode As String = "@CompanyCode"
-        Private Const pCompanyName As String = "@CompanyName"
-        Private Const pCompanySearchTerm1 As String = "@CompanySearchTerm1"
-        Private Const pCompanySearchTerm2 As String = "@CompanySearchTerm2"
-        Private Const pCompanyDescription As String = "@CompanyDescription"
+        ' This Module Identifier
+        Private varThisModuleId As Long = 0
+        Private Const varThisModuleCode As String = "CCIN"
 #End Region
 
 #Region "Subs Collections"
-        <SupportedOSPlatform("windows")>
+        <System.Runtime.Versioning.SupportedOSPlatform("windows")>
         Private Sub CheckAllInput()
-            'Call CheckAllControls()
             TxtCode.Focus()
             TxtName.Focus()
             TxtSearchTerm1.Focus()
@@ -31,7 +23,7 @@ Namespace UI.Canvas
 
 #Region "Form Events"
 
-        <SupportedOSPlatform("windows")>
+        <System.Runtime.Versioning.SupportedOSPlatform("windows")>
         Private Sub FRMccinEditor_Load(sender As Object, e As EventArgs) Handles MyBase.Load
             If (varDataProperties.CompanyIsNew) Then
                 ChkAddNew.Visible = True
@@ -40,13 +32,13 @@ Namespace UI.Canvas
                 ChkAddNew.Visible = False
                 ChkAddNew.Checked = False
                 CMDccin.Editor.GetCompanyProperties(varDataProperties, varDatasetIngrid)
-                If varDatasetIngrid.Tables(tCompany).Rows.Count > 0 Then
-                    With varDatasetIngrid.Tables(tCompany).Rows(0)
-                        TxtCode.Text = .Item("company_code").ToString
-                        TxtName.Text = .Item("company_name").ToString
-                        TxtSearchTerm1.Text = .Item("company_searchterm1").ToString
-                        TxtSearchTerm2.Text = .Item("company_searchterm2").ToString
-                        TxtDescription.Text = .Item("company_description").ToString
+                If varDatasetIngrid.Tables(tCompany.TableName).Rows.Count > 0 Then
+                    With varDatasetIngrid.Tables(tCompany.TableName).Rows(0)
+                        TxtCode.Text = .Item(tCompany.C_CompanyCode).ToString
+                        TxtName.Text = .Item(tCompany.C_CompanyName).ToString
+                        TxtSearchTerm1.Text = .Item(tCompany.C_SearchTerm1).ToString
+                        TxtSearchTerm2.Text = .Item(tCompany.C_SearchTerm2).ToString
+                        TxtDescription.Text = .Item(tCompany.C_Description).ToString
                     End With
                 End If
             End If
@@ -54,29 +46,24 @@ Namespace UI.Canvas
 #End Region
 
 #Region "Component Events"
-        <SupportedOSPlatform("windows")>
+        <System.Runtime.Versioning.SupportedOSPlatform("windows")>
         Private Sub BtnCancel_Click(sender As Object, e As EventArgs) Handles BtnCancel.Click
             Me.Close()
         End Sub
 
-        <SupportedOSPlatform("windows")>
+        <System.Runtime.Versioning.SupportedOSPlatform("windows")>
         Private Sub BtnSave_Click(sender As Object, e As EventArgs) Handles BtnSave.Click
             Call CheckAllInput()
 
             With varDataProperties
-                .AllParameters.Remove(pCompanyCode)
-                .AllParameters.Add(pCompanyCode, TxtCode.XOSqlText)
-                .AllParameters.Remove(pCompanyName)
-                .AllParameters.Add(pCompanyName, TxtName.XOSqlText)
-                .AllParameters.Remove(pCompanySearchTerm1)
-                .AllParameters.Add(pCompanySearchTerm1, IIf(TxtSearchTerm1.XOSqlText = String.Empty OrElse TxtSearchTerm1.XOSqlText = "", DBNull.Value, TxtSearchTerm1.XOSqlText))
-                .AllParameters.Remove(pCompanySearchTerm2)
-                .AllParameters.Add(pCompanySearchTerm2, IIf(TxtSearchTerm2.XOSqlText = String.Empty OrElse TxtSearchTerm2.XOSqlText = "", DBNull.Value, TxtSearchTerm2.XOSqlText))
-                .AllParameters.Remove(pCompanyDescription)
-                .AllParameters.Add(pCompanyDescription, IIf(TxtDescription.XOSqlText = String.Empty OrElse TxtDescription.XOSqlText = "", DBNull.Value, TxtDescription.XOSqlText))
+                SetValue(.AllParameters, tCompany.P_CompanyCode, TxtCode.XOSqlText)
+                SetValue(.AllParameters, tCompany.P_CompanyName, TxtName.XOSqlText)
+                SetValue(.AllParameters, tCompany.P_SearchTerm1, IIf(TxtSearchTerm1.XOSqlText = String.Empty OrElse TxtSearchTerm1.XOSqlText = "", DBNull.Value, TxtSearchTerm1.XOSqlText))
+                SetValue(.AllParameters, tCompany.P_SearchTerm2, IIf(TxtSearchTerm2.XOSqlText = String.Empty OrElse TxtSearchTerm2.XOSqlText = "", DBNull.Value, TxtSearchTerm2.XOSqlText))
+                SetValue(.AllParameters, tCompany.P_Description, IIf(TxtDescription.XOSqlText = String.Empty OrElse TxtDescription.XOSqlText = "", DBNull.Value, TxtDescription.XOSqlText))
             End With
 
-            If (varDataProperties.AllParameters(pCompanyCode).ToString = String.Empty) OrElse (varDataProperties.AllParameters(pCompanyName).ToString = String.Empty) Then
+            If (varDataProperties.AllParameters(tCompany.P_CompanyCode).ToString = String.Empty) OrElse (varDataProperties.AllParameters(tCompany.P_CompanyName).ToString = String.Empty) Then
                 Decision(My.Application.Info.AssemblyName.ToUpper, "Cannot save your record." & Environment.NewLine & "Ensure that the Company Code and Company Name fields are properly filled in.", LibApp.Ingrid.Global.PopupType.Alert, "", CMCv.UI.Canvas.FRMdialogbox.MessageIcon.Alert, CMCv.UI.Canvas.FRMdialogbox.MessageTypes.OkOnly)
                 Call ClearParameters()
                 Return
@@ -111,14 +98,25 @@ Namespace UI.Canvas
             End If
         End Sub
 
+        ''' <summary>
+        ''' 
+        ''' </summary>
+        <System.Runtime.Versioning.SupportedOSPlatform("windows")>
         Private Sub ClearParameters()
             With varDataProperties
-                .AllParameters.Remove("@CompanyCode")
-                .AllParameters.Remove("@CompanyName")
-                .AllParameters.Remove("@CompanySearchTerm1")
-                .AllParameters.Remove("@CompanySearchTerm2")
-                .AllParameters.Remove("@CompanyDescription")
+                .AllParameters.Remove(tCompany.P_CompanyCode)
+                .AllParameters.Remove(tCompany.P_CompanyName)
+                .AllParameters.Remove(tCompany.P_SearchTerm1)
+                .AllParameters.Remove(tCompany.P_SearchTerm2)
+                .AllParameters.Remove(tCompany.P_Description)
             End With
+        End Sub
+
+        <System.Runtime.Versioning.SupportedOSPlatform("windows")>
+        Private Sub FRMccinEditor_Activated(sender As Object, e As EventArgs) Handles Me.Activated
+            ' Set active module to UserParameters
+            SetValue(varDataProperties.UserParameters, tModule.P_ModuleId, varThisModuleId)
+            SetValue(varDataProperties.UserParameters, tModule.P_ModuleCode, varThisModuleCode)
         End Sub
 #End Region
