+++ Apps/Core/Vb/UI/Dataset/ADSglobal.xss
@@ -4,18 +4,19 @@
      Changes to this file may cause incorrect behavior and will be lost if
      the code is regenerated.
 </autogenerated>-->
-<DiagramLayout xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" ex:showrelationlabel="False" ViewPortX="-220" ViewPortY="226" xmlns:ex="urn:schemas-microsoft-com:xml-msdatasource-layout-extended" xmlns="urn:schemas-microsoft-com:xml-msdatasource-layout">
+<DiagramLayout xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" ex:showrelationlabel="False" ViewPortX="-181" ViewPortY="48" xmlns:ex="urn:schemas-microsoft-com:xml-msdatasource-layout-extended" xmlns="urn:schemas-microsoft-com:xml-msdatasource-layout">
   <Shapes>
-    <Shape ID="DesignTable:DatabaseProperties" ZOrder="2" X="112" Y="58" Height="118" Width="245" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="114" />
-    <Shape ID="DesignTable:man_company" ZOrder="9" X="542" Y="63" Height="199" Width="241" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="195" />
-    <Shape ID="DesignTable:CheckRelation" ZOrder="1" X="-90" Y="62" Height="307" Width="177" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="303" />
-    <Shape ID="DesignTable:man_department" ZOrder="10" X="543" Y="287" Height="172" Width="237" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="168" />
-    <Shape ID="DesignTable:man_position" ZOrder="8" X="546" Y="490" Height="226" Width="237" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="222" />
-    <Shape ID="DesignTable:mat_plant" ZOrder="7" X="812" Y="71" Height="307" Width="207" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="303" />
-    <Shape ID="DesignTable:UserData" ZOrder="6" X="111" Y="212" Height="307" Width="293" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="303" />
-    <Shape ID="DesignTable:EPLS_Editor" ZOrder="5" X="-184" Y="398" Height="605" Width="267" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="601" />
-    <Shape ID="DesignTable:SysModule" ZOrder="4" X="-182" Y="1029" Height="199" Width="262" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="195" />
-    <Shape ID="DesignTable:SYSS_Editor" ZOrder="3" X="114" Y="558" Height="497" Width="289" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="493" />
+    <Shape ID="DesignTable:DatabaseProperties" ZOrder="3" X="112" Y="58" Height="118" Width="245" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="114" />
+    <Shape ID="DesignTable:man_company" ZOrder="10" X="665" Y="71" Height="199" Width="241" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="195" />
+    <Shape ID="DesignTable:CheckRelation" ZOrder="2" X="-90" Y="62" Height="307" Width="177" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="303" />
+    <Shape ID="DesignTable:man_department" ZOrder="11" X="667" Y="285" Height="172" Width="237" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="168" />
+    <Shape ID="DesignTable:man_position" ZOrder="9" X="664" Y="473" Height="226" Width="237" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="222" />
+    <Shape ID="DesignTable:mat_plant" ZOrder="8" X="917" Y="67" Height="307" Width="207" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="303" />
+    <Shape ID="DesignTable:UserData" ZOrder="7" X="111" Y="212" Height="307" Width="293" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="303" />
+    <Shape ID="DesignTable:EPLS_Editor" ZOrder="6" X="-184" Y="398" Height="605" Width="267" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="601" />
+    <Shape ID="DesignTable:SysModule" ZOrder="5" X="-182" Y="1029" Height="199" Width="262" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="195" />
+    <Shape ID="DesignTable:SYSS_Editor" ZOrder="4" X="114" Y="558" Height="551" Width="289" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="547" />
+    <Shape ID="DesignTable:file_attachment" ZOrder="1" X="918" Y="395" Height="65" Width="214" AdapterExpanded="true" DataTableExpanded="true" OldAdapterHeight="0" OldDataTableHeight="0" SplitterPosition="61" />
   </Shapes>
   <Connectors />
 </DiagramLayout>
